A smorgasborg of Changes
Altered physics constants for a more satifying experience
Added a wall using the ground class
The player class now handles projectile creation
Player class now accepts a list of rectangles. Currently, this is a list of all the "ground" rectangles
Projectile class now gets the same list from the player. There might be a better way to do this, but it's where we are now.
Player class has gravity again (but does nothing with it for now)
Parentheses run amok