A mass of changes
Adjusted the physics of the game so that the game runs at an understandable speed. This is done by reducing gravity, power of shots, and jump velocity.

Player once again have a power property. This is passed to the explosive class upon explosive creation.

Player once again has the ability to change the angle of the shots based on the mouse pointer location.

Lowered the movement speed of the players.

Players no longer can change velocity in mid-air.